+++ src/providers/seventv/SeventvEventAPI.cpp
@@ -1,9 +1,9 @@
 #include "providers/seventv/SeventvEventAPI.hpp"
 
 #include "Application.hpp"
-#include "providers/seventv/eventapi/SeventvEventAPIClient.hpp"
-#include "providers/seventv/eventapi/SeventvEventAPIDispatch.hpp"
-#include "providers/seventv/eventapi/SeventvEventAPIMessage.hpp"
+#include "providers/seventv/eventapi/Client.hpp"
+#include "providers/seventv/eventapi/Dispatch.hpp"
+#include "providers/seventv/eventapi/Message.hpp"
 #include "providers/seventv/SeventvBadges.hpp"
 #include "providers/seventv/SeventvCosmetics.hpp"
 #include "providers/seventv/SeventvPaints.hpp"
