+++ src/providers/seventv/eventapi/Dispatch.cpp
@@ -1,4 +1,4 @@
-#include "providers/seventv/eventapi/SeventvEventAPIDispatch.hpp"
+#include "providers/seventv/eventapi/Dispatch.hpp"
 
 #include <utility>
