+++ tests/src/SeventvEventAPI.cpp
@@ -1,14 +1,15 @@
 #include "providers/seventv/SeventvEventAPI.hpp"
 
-#include "providers/seventv/eventapi/SeventvEventAPIClient.hpp"
-#include "providers/seventv/eventapi/SeventvEventAPIDispatch.hpp"
-#include "providers/seventv/eventapi/SeventvEventAPIMessage.hpp"
+#include "providers/seventv/eventapi/Client.hpp"
+#include "providers/seventv/eventapi/Dispatch.hpp"
+#include "providers/seventv/eventapi/Message.hpp"
 
 #include <boost/optional.hpp>
 #include <gtest/gtest.h>
 #include <QString>
 
 using namespace chatterino;
+using namespace chatterino::seventv::eventapi;
 using namespace std::chrono_literals;
 
 const QString EMOTE_SET_A = "60b39e943e203cc169dfc106";
@@ -21,10 +22,10 @@ TEST(SeventvEventAPI, AllEvents)
     auto *eventAPI = new SeventvEventAPI(host, std::chrono::milliseconds(1000));
     eventAPI->start();
 
-    boost::optional<SeventvEventAPIEmoteAddDispatch> addDispatch;
-    boost::optional<SeventvEventAPIEmoteUpdateDispatch> updateDispatch;
-    boost::optional<SeventvEventAPIEmoteRemoveDispatch> removeDispatch;
-    boost::optional<SeventvEventAPIUserConnectionUpdateDispatch> userDispatch;
+    boost::optional<EmoteAddDispatch> addDispatch;
+    boost::optional<EmoteUpdateDispatch> updateDispatch;
+    boost::optional<EmoteRemoveDispatch> removeDispatch;
+    boost::optional<UserConnectionUpdateDispatch> userDispatch;
 
     eventAPI->signals_.emoteAdded.connect([&](const auto &d) {
         addDispatch = d;
